뉴스 게시판과 관리자 알림 센터
뉴스 게시판은 공지사항 게시판을 갖다 붙였고,  알림 센터는 게시판처럼 보여주면 DTO를 5개나 가져와야 하는 부담감에 겉 모습을 개편하여 관리자 메인 화면으로 쓰였습니다.
