+++ SsangYongBang/WebContent/WEB-INF/views/inc/header.jsp
@@ -26,7 +26,7 @@
           			<a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-expanded="false">커뮤니티<span class="caret"></span></a>
 			          	<ul class="dropdown-menu" role="menu">
 				            <li><a href="/sybang/notice/boardList.do">공지사항</a></li>
-				            <li><a href="#">뉴스</a></li>
+				            <li><a href="/sybang/news/boardList.do">뉴스</a></li>
 				            <li><a href="/sybang/board/freelist.do">자유게시판</a></li>
 				            <li><a href="#">문의게시판</a></li>
 			         	</ul>
