+++ SsangYongBang/src/com/test/sist/inc/Footer.java
@@ -1,0 +1,23 @@
+package com.test.sist.inc;
+
+import java.io.IOException;
+
+import javax.servlet.RequestDispatcher;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@WebServlet("/inc/footer.do")
+public class Footer extends HttpServlet {
+
+	protected void doGet(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
+
+		RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/views/inc/footer.jsp");
+		dispatcher.forward(request, response);
+
+	}
+
+}
