+++ SsangYongBang/src/com/test/sist/news/BoardList.java
@@ -1,4 +1,4 @@
-package com.test.sist.notice;
+package com.test.sist.news;
 
 import java.io.IOException;
 
@@ -9,13 +9,13 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-@WebServlet("/inc/header.do")
-public class Header extends HttpServlet {
+@WebServlet("/news/boardList.do")
+public class BoardList extends HttpServlet {
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 
-		RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/views/inc/header.jsp");
+		RequestDispatcher dispatcher = request.getRequestDispatcher("/WEB-INF/views/news/boardList.jsp");
 		dispatcher.forward(request, response);
 
 	}
